requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alabaster==0.7.13
anyio==3.7.1
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
Automat==22.10.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.0.12
cheroot==10.0.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
constantly==15.1.0
contourpy==1.1.0
converter==1.0.0
coverage==7.3.0
cryptography==41.0.3
cssselect==1.2.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2.5
django-cache-machine==1.2.0
django-redis==5.3.0
dnspython==2.4.2
docutils==0.20.1
executing==1.2.0
fake-useragent==1.2.1
fastjsonschema==2.18.0
filelock==3.12.3
fire==0.5.0
flake8==6.1.0
Flask==3.0.0
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
geoip2==4.7.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
imagesize==1.4.1
incremental==22.10.0
iniconfig==2.0.0
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
items==0.6.5
itsdangerous==2.1.2
jaraco.functools==4.0.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
maxminddb==2.4.0
mccabe==0.7.0
mistune==3.0.1
mock==5.1.0
more-itertools==10.1.0
multidict==6.0.4
mysql-connector-python==8.1.0
mysqlclient==2.2.0
nbclient==0.8.0
nbconvert==7.7.4
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.0.2
notebook_shim==0.2.3
ntlm-auth==1.5.0
ntplib==0.4.0
nulltype==2.3.1
numpy==1.25.2
opencv-python==4.8.0.76
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
parsel==1.8.1
parso==0.8.3
pdf2docx==0.5.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
pluggy==1.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
Protego==0.3.0
protobuf==4.21.12
psutil==5.9.5
psycopg==3.1.10
psycopg-binary==3.1.10
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
PyDispatcher==2.0.7
pyfiglet==0.8.post1
pyflakes==3.1.0
pygame==2.5.1
Pygments==2.16.1
pymongo==4.5.0
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
PyMySQL==1.1.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyproject-api==1.6.1
PySocks==1.7.1
pyspnego==0.9.2
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.8.2
python-docx==0.8.11
python-json-logger==2.0.7
python-memcached==1.59
python-nmap==0.7.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.3
QtPy==2.3.1
queuelib==1.6.2
redis==5.0.0
referencing==0.30.2
requests==2.31.0
requests-file==1.5.1
requests-ntlm==1.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.0
schedule==1.2.0
Scrapy==2.11.0
scrapy-redis==0.7.3
selenium==4.15.2
Send2Trash==1.8.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4.1
Sphinx==7.2.5
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sqlparse==0.4.4
stack-data==0.6.2
style==1.1.0
termcolor==2.3.0
terminado==0.17.1
tinycss2==1.2.1
tldextract==3.4.4
tornado==6.3.3
tox==4.11.1
tox-gh-actions==3.1.3
traitlets==5.9.0
trio==0.23.1
trio-websocket==0.11.1
Twisted==22.10.0
typing_extensions==4.7.1
tzdata==2023.3
update==0.0.1
uri-template==1.3.0
urllib3==2.0.4
uvloop==0.17.0
virtualenv==20.24.4
virustotal-python==1.0.2
w3lib==2.1.2
wcwidth==0.2.6
web.py==0.62
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
websockets==11.0.3
Werkzeug==3.0.0
widgetsnbextension==4.0.8
wsproto==1.2.0
yarl==1.9.2
zope.interface==6.0
